resource

+++ WinApiLecture/Client/Struct.h
@@ -9,7 +9,7 @@ struct Vec2
 public:
 	float Length()
 	{
-		return sqrt(x * x + y * y);
+		return (float)sqrt(x * x + y * y);
 	}
 	Vec2& Normalize()
 	{
